Simplification & Fixes
